+++ numpy/core/include/numpy/random/bitgen.h
