+++ numpy/random/_generator.pyx
@@ -28,7 +28,7 @@ from ._common cimport (POISSON_LAM_MAX, CONS_POSITIVE, CONS_NONE,
         )
 
 
-cdef extern from "include/distributions.h":
+cdef extern from "numpy/random/distributions.h":
 
     struct s_binomial_t:
         int has_binomial
